Reorganizing the test data
